refactor: theme, time의 admin만 가능한 로직들의 POST, DELETE api를 admin으로 시작하도록 변경
