+++ src/main/java/roomescape/controller/theme/AdminThemeController.java
@@ -1,0 +1,44 @@
+package roomescape.controller.theme;
+
+import jakarta.validation.Valid;
+import java.net.URI;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.DeleteMapping;
+import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
+import org.springframework.web.util.UriComponentsBuilder;
+import roomescape.controller.theme.dto.CreateThemeRequest;
+import roomescape.controller.theme.dto.ThemeResponse;
+import roomescape.service.ThemeService;
+
+@RestController
+@RequestMapping("/admin/themes")
+public class AdminThemeController {
+    private final ThemeService themeService;
+
+    public AdminThemeController(final ThemeService themeService) {
+        this.themeService = themeService;
+    }
+
+    @PostMapping
+    public ResponseEntity<ThemeResponse> addTheme(
+            @RequestBody @Valid final CreateThemeRequest createThemeRequest) {
+        final ThemeResponse theme = themeService.addTheme(createThemeRequest);
+
+        final URI uri = UriComponentsBuilder.fromPath("/admin/themes/{id}")
+                .buildAndExpand(theme.id())
+                .toUri();
+        return ResponseEntity.created(uri)
+                .body(theme);
+    }
+
+    @DeleteMapping("/{id}")
+    public ResponseEntity<Void> deleteTheme(@PathVariable("id") final long id) {
+        themeService.deleteTheme(id);
+        return ResponseEntity.noContent()
+                .build();
+    }
+}
