+++ src/main/java/roomescape/controller/theme/ThemeController.java
@@ -1,19 +1,11 @@
 package roomescape.controller.theme;
 
 import jakarta.validation.Valid;
-import java.net.URI;
 import java.util.List;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
-import org.springframework.web.util.UriComponentsBuilder;
 import roomescape.controller.reservation.dto.PopularThemeResponse;
-import roomescape.controller.theme.dto.CreateThemeRequest;
 import roomescape.controller.theme.dto.PopularThemeRequest;
 import roomescape.controller.theme.dto.ThemeResponse;
 import roomescape.service.ThemeService;
@@ -31,25 +23,6 @@ public class ThemeController {
     @GetMapping
     public List<ThemeResponse> getThemes() {
         return themeService.getThemes();
-    }
-
-    @PostMapping
-    public ResponseEntity<ThemeResponse> addTheme(
-            @RequestBody @Valid final CreateThemeRequest createThemeRequest) {
-        final ThemeResponse theme = themeService.addTheme(createThemeRequest);
-
-        final URI uri = UriComponentsBuilder.fromPath("/themes/{id}")
-                .buildAndExpand(theme.id())
-                .toUri();
-        return ResponseEntity.created(uri)
-                .body(theme);
-    }
-
-    @DeleteMapping("/{id}")
-    public ResponseEntity<Void> deleteTheme(@PathVariable("id") final long id) {
-        themeService.deleteTheme(id);
-        return ResponseEntity.noContent()
-                .build();
     }
 
     @GetMapping(value = "/popular", params = {"from", "until", "limit"})
