+++ src/test/java/roomescape/controller/theme/AdminThemeControllerTest.java
@@ -1,0 +1,132 @@
+package roomescape.controller.theme;
+
+import io.restassured.RestAssured;
+import io.restassured.http.ContentType;
+import java.util.HashMap;
+import java.util.Map;
+import java.util.stream.Stream;
+import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.DisplayName;
+import org.junit.jupiter.api.Test;
+import org.junit.jupiter.params.ParameterizedTest;
+import org.junit.jupiter.params.provider.Arguments;
+import org.junit.jupiter.params.provider.MethodSource;
+import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.boot.test.web.server.LocalServerPort;
+import org.springframework.test.context.jdbc.Sql;
+import org.springframework.test.context.jdbc.Sql.ExecutionPhase;
+import roomescape.controller.member.dto.MemberLoginRequest;
+
+@SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT)
+@Sql(value = "/data.sql", executionPhase = ExecutionPhase.AFTER_TEST_METHOD)
+public class AdminThemeControllerTest {
+
+    @LocalServerPort
+    int port;
+
+    String adminToken;
+    String memberToken;
+
+    @BeforeEach
+    void setUp() {
+        RestAssured.port = port;
+
+        adminToken = RestAssured
+                .given().log().all()
+                .body(new MemberLoginRequest("redddy@gmail.com", "0000"))
+                .contentType(ContentType.JSON)
+                .when().post("/login")
+                .then().log().cookies().extract().cookie("token");
+
+        memberToken = RestAssured
+                .given().log().all()
+                .body(new MemberLoginRequest("jinwuo0925@gmail.com", "1111"))
+                .contentType(ContentType.JSON)
+                .when().post("/login")
+                .then().log().cookies().extract().cookie("token");
+    }
+
+    @Test
+    @DisplayName("테마 생성")
+    void addTheme() {
+        final Map<String, String> params = new HashMap<>();
+        params.put("name", "테마 이름");
+        params.put("description", "테마 설명");
+        params.put("thumbnail", "https://google.png");
+
+        RestAssured.given().log().all()
+                .cookie("token", adminToken)
+                .contentType(ContentType.JSON)
+                .body(params)
+                .when().post("/admin/themes")
+                .then().log().all()
+                .statusCode(201);
+    }
+
+    @Test
+    @DisplayName("예약된 테마는 삭제할 수 없다.")
+    void deleteTheme() {
+        RestAssured.given().log().all()
+                .cookie("token", adminToken)
+                .when().delete("/admin/themes/1")
+                .then().log().all()
+                .statusCode(400);
+    }
+
+    static Stream<Arguments> invalidRequestParameterProvider() {
+        final String name = "name";
+        final String description = "description";
+        final String thumbnail = "thumbnail";
+        final String over255 = "1".repeat(256);
+        return Stream.of(
+                Arguments.of(name, description, null),
+                Arguments.of(name, null, thumbnail),
+                Arguments.of(null, description, thumbnail),
+                Arguments.of(over255, description, thumbnail),
+                Arguments.of(name, over255, thumbnail),
+                Arguments.of(name, description, over255)
+        );
+    }
+
+    @ParameterizedTest
+    @MethodSource("invalidRequestParameterProvider")
+    @DisplayName("유효하지 않은 요청인 경우 400을 반환한다.")
+    void invalidRequest(final String name, final String description, final String thumbnail) {
+        final Map<String, String> params = new HashMap<>();
+        params.put("name", name);
+        params.put("description", description);
+        params.put("thumbnail", thumbnail);
+
+        RestAssured.given().log().all()
+                .cookie("token", adminToken)
+                .contentType(ContentType.JSON)
+                .body(params)
+                .when().post("/admin/themes")
+                .then().log().all()
+                .statusCode(400);
+    }
+
+    @Test
+    @DisplayName("권한이 없는 경우 401을 응답한다.")
+    void accessDenied() {
+        final Map<String, String> params = new HashMap<>();
+        params.put("name", "테마 이름");
+        params.put("description", "테마 설명");
+        params.put("thumbnail", "https://google.png");
+
+        RestAssured.given().log().all()
+                .cookie("token", memberToken)
+                .contentType(ContentType.JSON)
+                .body(params)
+                .when().post("/admin/themes")
+                .then().log().all()
+                .statusCode(401);
+
+        RestAssured.given().log().all()
+                .cookie("token", memberToken)
+                .contentType(ContentType.JSON)
+                .when().delete("/admin/themes/1")
+                .then().log().all()
+                .statusCode(401);
+    }
+}
