+++ src/test/java/roomescape/controller/time/TimeControllerTest.java
@@ -3,19 +3,11 @@ package roomescape.controller.time;
 import static org.hamcrest.Matchers.is;
 
 import io.restassured.RestAssured;
-import io.restassured.http.ContentType;
 import java.time.LocalDate;
 import java.time.format.DateTimeFormatter;
-import java.util.HashMap;
-import java.util.List;
-import java.util.Map;
-import java.util.stream.Stream;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import org.junit.jupiter.params.ParameterizedTest;
-import org.junit.jupiter.params.provider.Arguments;
-import org.junit.jupiter.params.provider.MethodSource;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.boot.test.web.server.LocalServerPort;
 import org.springframework.http.MediaType;
@@ -30,15 +22,15 @@ class TimeControllerTest {
     @LocalServerPort
     int port;
 
-    String accessToken;
+    String memberToken;
 
     @BeforeEach
     void setUp() {
         RestAssured.port = port;
 
-        accessToken = RestAssured
+        memberToken = RestAssured
                 .given().log().all()
-                .body(new MemberLoginRequest("redddy@gmail.com", "0000"))
+                .body(new MemberLoginRequest("jinwuo0925@gmail.com", "1111"))
                 .contentType(MediaType.APPLICATION_JSON_VALUE)
                 .when().post("/login")
                 .then().log().cookies().extract().cookie("token");
@@ -48,73 +40,11 @@ class TimeControllerTest {
     @DisplayName("타임 조회")
     void getTimes() {
         RestAssured.given().log().all()
-                .cookie("token", accessToken)
+                .cookie("token", memberToken)
                 .when().get("/times")
                 .then().log().all()
                 .statusCode(200)
                 .body("size()", is(5));
-    }
-
-    static Stream<Arguments> invalidRequestParameterProvider() {
-        return Stream.of(
-                Arguments.of("20202020"),
-                Arguments.of("1234"),
-                Arguments.of("2026-"),
-                Arguments.of("2026-13-01"),
-                Arguments.of("2026-12-32"),
-                Arguments.of("2026-11"),
-                Arguments.of("2026-1")
-        );
-    }
-
-    @ParameterizedTest
-    @MethodSource("invalidRequestParameterProvider")
-    @DisplayName("유효하지 않은 요청인 경우 400을 반환한다.")
-    void invalidRequest(final String startAt) {
-        final Map<String, String> params = new HashMap<>();
-        params.put("startAt", startAt);
-
-        RestAssured.given().log().all()
-                .cookie("token", accessToken)
-                .contentType(ContentType.JSON)
-                .body(params)
-                .when().post("/times")
-                .then().log().all()
-                .statusCode(400);
-    }
-
-    @Test
-    @DisplayName("시간 삭제")
-    void deleteTime() {
-        Map<String, String> params = new HashMap<>();
-        params.put("startAt", "00:00");
-
-        RestAssured.given().log().all()
-                .cookie("token", accessToken)
-                .contentType(ContentType.JSON)
-                .body(params)
-                .when().post("/times")
-                .then().log().all()
-                .statusCode(201);
-
-        final List<Object> values = RestAssured.given().log().all()
-                .cookie("token", accessToken)
-                .when().get("/times")
-                .then().log().all()
-                .statusCode(200)
-                .extract().jsonPath().getList("$");
-
-        RestAssured.given().log().all()
-                .cookie("token", accessToken)
-                .when().delete("/times/" + values.size())
-                .then().log().all()
-                .statusCode(204);
-
-        RestAssured.given().log().all()
-                .cookie("token", accessToken)
-                .when().delete("/times/" + values.size())
-                .then().log().all()
-                .statusCode(400);
     }
 
     @Test
@@ -124,7 +54,7 @@ class TimeControllerTest {
         long themeId = 1L;
 
         RestAssured.given().log().all()
-                .cookie("token", accessToken)
+                .cookie("token", memberToken)
                 .when().get("/times/availability?date=" + date.format(DateTimeFormatter.ISO_DATE)
                         + "&themeId=" + themeId)
                 .then().log().all()
